+++ lib/domain/menstruation/menstruation_store.dart
@@ -13,7 +13,7 @@ import 'package:pilll/database/pill_sheet.dart';
 import 'package:pilll/database/pill_sheet_group.dart';
 import 'package:pilll/database/setting.dart';
 import 'package:pilll/domain/menstruation/menstruation_state.codegen.dart';
-import 'package:pilll/service/user.dart';
+import 'package:pilll/database/user.dart';
 import 'package:pilll/util/datetime/day.dart';
 
 final menstruationsStoreProvider =
