+++ lib/domain/menstruation_edit/menstruation_edit_store.dart
@@ -9,7 +9,7 @@ import 'package:pilll/native/health_care.dart';
 import 'package:pilll/database/menstruation.dart';
 import 'package:pilll/database/setting.dart';
 import 'package:pilll/domain/menstruation_edit/menstruation_edit_state.codegen.dart';
-import 'package:pilll/service/user.dart';
+import 'package:pilll/database/user.dart';
 import 'package:pilll/util/datetime/date_compare.dart';
 import 'package:pilll/util/datetime/day.dart';
