+++ lib/domain/premium_trial/premium_trial_modal_store.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:pilll/domain/premium_trial/premium_trial_modal_state.codegen.dart';
 import 'package:pilll/entity/user_error.dart';
 import 'package:pilll/error_log.dart';
-import 'package:pilll/service/user.dart';
+import 'package:pilll/database/user.dart';
 import 'package:riverpod/riverpod.dart';
 
 final premiumTrialStoreProvider =
